refactor(auth): move FormField into auth feature
