+++ frontend/src/pages/Login.tsx
@@ -2,7 +2,7 @@ import { useState } from "react";
 import { useApiError } from "../hooks/useApiError";
 import { Link, useNavigate } from "react-router-dom"
 import { AuthCard } from "../features/auth/components/AuthCard";
-import { FormField } from "../components/FormField";
+import { FormField } from "../features/auth/components/FormField";
 import { login } from "../lib/api";
 
 export const Login = () => {
